Take FileConfig out of CFamilyHelper partial
